requirements.txt
absl-py==2.1.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.15.0
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
bullet-safety-gym==1.4.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
docker-pycreds==0.4.0
dsrl==0.1.0
etils==1.7.0
exceptiongroup==1.2.1
executing==2.0.1
farama-notifications==0.0.4
fastjsonschema==2.19.1
flax==0.8.3
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
gitpython==3.1.43
glfw==2.7.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
importlib-resources==6.4.0
ipykernel==6.29.4
ipython==8.24.0
isoduration==20.11.0
isort==5.13.2
jax==0.4.28
jax-cuda12-pjrt==0.4.28
jax-cuda12-plugin==0.4.28
jax-jumpy==1.0.0
jaxlib==0.4.28
jedi==0.19.1
jinja2==3.1.4
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.1
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.0
jupyter-server-terminals==0.5.3
jupyterlab==4.2.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-collections==0.1.1
ml-dtypes==0.4.0
msgpack==1.0.8
mujoco==2.3.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook-shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.4.131
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.2.1.3
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvitop==1.3.2
opt-einsum==3.3.0
optax==0.2.2
orbax-checkpoint==0.5.11
overrides==7.7.0
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
pip==24.0
platformdirs==4.2.1
prometheus-client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pybullet==3.2.6
pycparser==2.22
pygame==2.1.0
pygments==2.18.0
pyopengl==3.1.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pyyaml==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.4
safety-gymnasium==0.4.0
scipy==1.13.0
send2trash==1.8.3
sentry-sdk==2.1.1
setproctitle==1.3.3
setuptools==65.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
tensorflow-probability==0.24.0
tensorstore==0.1.59
termcolor==2.4.0
terminado==0.18.1
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing-extensions==4.11.0
uri-template==1.3.0
urllib3==2.2.1
uv==0.1.42
wandb==0.17.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.43.0
xmltodict==0.13.0
zipp==3.18.1
